refactor: move git pkg into runner/pkg/
