+++ runner/BUILD.bazel
@@ -9,9 +9,9 @@ go_library(
         "//common/bazel",
         "//language/js",
         "//language/orion",
-        "//runner/git",
         "//runner/language/bzl",
         "//runner/language/python",
+        "//runner/pkg/git",
         "//runner/pkg/ibp",
         "//runner/progress",
         "//runner/vendored/gazelle",
