+++ runner/pkg/git/testing/BUILD.bazel
@@ -4,10 +4,10 @@ go_library(
     name = "testing",
     testonly = True,
     srcs = ["language.go"],
-    importpath = "github.com/aspect-build/aspect-gazelle/runner/git/testing",
+    importpath = "github.com/aspect-build/aspect-gazelle/runner/pkg/git/testing",
     visibility = ["//visibility:public"],
     deps = [
-        "//runner/git",
+        "//runner/pkg/git",
         "@gazelle//config",
         "@gazelle//label",
         "@gazelle//language",
