+++ runner/pkg/git/tests/ignore_config_files/root.starstar-ig.ts
