+++ runner/pkg/git/tests/ignore_config_files/subdir/nested/x.nested-ig.ts
