+++ runner/pkg/git/tests/ignore_config_files/subdir/nested/y.starstar-ig.ts
