+++ runner/pkg/git/tests/ignore_config_files/subdir/star-ig.ts
