+++ runner/pkg/git/tests/ignore_config_files/subdir2/y.starstar-ig.ts
