+++ runner/runner.go
@@ -28,9 +28,9 @@ import (
 	"github.com/aspect-build/aspect-gazelle/common/bazel"
 	js "github.com/aspect-build/aspect-gazelle/language/js"
 	orion "github.com/aspect-build/aspect-gazelle/language/orion"
-	"github.com/aspect-build/aspect-gazelle/runner/git"
 	"github.com/aspect-build/aspect-gazelle/runner/language/bzl"
 	"github.com/aspect-build/aspect-gazelle/runner/language/python"
+	"github.com/aspect-build/aspect-gazelle/runner/pkg/git"
 	"github.com/aspect-build/aspect-gazelle/runner/pkg/ibp"
 	"github.com/aspect-build/aspect-gazelle/runner/progress"
 	vendoredGazelle "github.com/aspect-build/aspect-gazelle/runner/vendored/gazelle"
